add fallback option to gcpsecrets provider
Using this with Helm and dynamic secret keys, and in some environments we want a fixed value, or we want the value to empty. The fixed value could be solved by adding a secret with that fixed value, but an empty value is not allowed in GCP.

By providing a fallback option an empty string can be set as fallback.