Add an option to prevent failing on missing key in map
When you specify the fragment part of the vals ref URI to let it dig the secret value encoded as json or yaml, it has been returning an error on missing key. You can now toggle the behavior so that it can return an empty value on missing key instead.

This might allow external tools like Helmfile to more easily fallback to a default value on a missing key error.

Note though, this does not change the behavior of the missing secret error. The missing secret error should not be handled in that way because I consider most of those secret management systems do not distinguish missing secret errors vs access forbidden errors.

Ref #52

Signed-off-by: Yusuke Kuoka <ykuoka@gmail.com>