support AWS_ENDPOINT_URL configuration
the AWS SDK and CLI (currently only in Python) allows for
configuration of the AWS endpoint URL via the `AWS_ENDPOINT_URL`
environment variable [1].

the Golang AWS SDK does not currently support this, but according to
GitHub issues, parity is on the way [2].

this change allows for a temporary workaround (it took two years for
this support to make it to the Python SDK [3]) that allows `vals` to
pick up the `AWS_ENDPOINT_URL` environment variable and configure the
session accordingly.

this allows for testing with non-AWS infrastructure like Localstack
and Moto, and could even open the door for more comprehensive
integration tests of the AWS functionality in this project without the
need for using real AWS infrastructure.

[1]: https://docs.aws.amazon.com/sdkref/latest/guide/feature-ss-endpoints.html
[2]: https://github.com/aws/aws-sdk-go/issues/4942
[3]: https://github.com/boto/boto3/pull/2746

Signed-off-by: Stephen Delano <steviedu@gmail.com>